+++ changelog.md
@@ -2,3 +2,9 @@
 
 ### Version 0.1-BETA
 * initial release
+
+### Version 0.2-BETA
+* Introduced inventory
+* Introduced Unlocking
+* converted element count to Decimal()
+* Added Progress bars to Production